test: stage2 해결 - 인터페이스 구현 책임을 팩토리에게 위임
